+++ chell/app/routes/api/courses/index.js
@@ -3,41 +3,11 @@ const mongoose = require('mongoose');
 const _ = require('lodash');
 
 const versionsRouter = require('./versions');
+const getCollectionFilter = require('../filters').getCollectionFilter;
+const getObjectFilter = require('../filters').getObjectFilter;
 const loadCourse = require('../../../middleware/load_course');
 
 const router = express.Router();
-
-//
-// Filters
-// TODO: Factor filters out of here.
-//
-
-function getCoursesFilter(courses) {
-  const ignoredAttributes = ['change_log', 'long_description'];
-
-  return _.map(courses, (course) => {
-    let filteredCourses = course;
-
-    // Snake case keys
-    // NOTE: This will convert `_id` to `id`.
-    filteredCourses = _.deeply(_.mapKeys)(filteredCourses, (v, k) =>
-      _.snakeCase(k));
-
-    // Drop ignored attributes
-    filteredCourses = _.deeply(_.omitBy)(filteredCourses, (v, k) =>
-      ignoredAttributes.includes(k));
-
-    return filteredCourses;
-  });
-}
-
-function getCourseFilter(course) {
-  return _.deeply(_.mapKeys)(course, (v, k) => _.snakeCase(k));
-}
-
-//
-// Routes
-//
 
 // GET /api/courses
 router.get('/', (req, res) => {
@@ -45,7 +15,7 @@ router.get('/', (req, res) => {
 
   Course.find().exec()
     .then((docs) => {
-      const courses = getCoursesFilter(_.map(docs, doc => doc.toJSON()));
+      const courses = getCollectionFilter(_.map(docs, doc => doc.toJSON()));
       res.status(200).json(courses);
     })
     .catch(err =>
@@ -54,7 +24,7 @@ router.get('/', (req, res) => {
 
 // GET /api/courses/:courseId
 router.get('/:courseId', loadCourse, (req, res) => {
-  const course = getCourseFilter(req.course.toJSON());
+  const course = getObjectFilter(req.course.toJSON());
   res.status(200).json(course);
 });