+++ chell/app/routes/api/courses/versions/index.js
@@ -1,28 +1,16 @@
 const express = require('express');
 const _ = require('lodash');
-const debug = require('debug')('chell:api');
+
+const getCollectionFilter = require('../../filters').getCollectionFilter;
+const getObjectFilter = require('../../filters').getObjectFilter;
+const loadVersion = require('../../../../middleware/load_version');
 
 const router = express.Router();
 
-function loadVersion(req, res, next) {
-  const version = _.find(req.course.versions, v =>
-    v.id === req.params.versionId);
-
-  if (!version) {
-    debug(`Version '${req.params.versionId}' not found in course '${req.course.id}'`);
-
-    const notFound = new Error('Not Found');
-    notFound.status = 404;
-    return next(notFound);
-  }
-
-  req.version = version; // eslint-disable-line no-param-reassign
-  return next();
-}
-
 // GET /api/courses/:courseId/versions
 router.get('/', (req, res) => {
-  res.status(200).json(req.course.versions);
+  const versions = _.map(req.course.versions, doc => doc.toJSON());
+  res.status(200).json(getCollectionFilter(versions));
 });
 
 // GET /api/courses/:courseId/versions/:versionId
@@ -30,7 +18,7 @@ router.get('/:versionId', loadVersion, (req, res, next) => {
   res.status(200);
 
   res.format({
-    json: () => res.json(req.version),
+    json: () => res.json(getObjectFilter(req.version.toJSON())),
 
     'application/x-gzip': () => {
       req.app.locals.fileStorage.get(req.version.archiveFilename)